backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
img2pdf
qrcode
pdf2docx
python-pptx
PyPDF2
